Big Ol' Namespace Cleanup
