+++ app/models/spree/payment_decorator.rb
@@ -1,7 +1,7 @@
-module Spree
+module Spree::Retail
   module PaymentDecorator
     delegate :pos_order_id, to: :order
   end
 end
 
-Spree::Payment.prepend Spree::PaymentDecorator
+Spree::Payment.prepend Spree::Retail::PaymentDecorator
