+++ app/models/spree/refund_decorator.rb
@@ -1,7 +1,7 @@
-module Spree
+module Spree::Retail
   module RefundDecorator
     delegate :pos_order_id, to: :payment
   end
 end
 
-Spree::Refund.prepend Spree::RefundDecorator
+Spree::Refund.prepend Spree::Retail::RefundDecorator
