+++ lib/generators/solidus_retail/install/install_generator.rb
@@ -1,4 +1,4 @@
-module SolidusRetail
+module Spree::Retail
   module Generators
     class InstallGenerator < Rails::Generators::Base
       class_option :auto_run_migrations, type: :boolean, default: false
