--- lib/solidus_retail/engine.rb
@@ -1,27 +1,0 @@
-require 'shopify_api'
-
-module SolidusRetail
-  class Engine < Rails::Engine
-    require 'spree/core'
-
-    isolate_namespace Spree
-    engine_name 'solidus_retail'
-
-    # TODO(cab): To remove and instead priorize a configuration initializer
-    require 'dotenv-rails'
-    Dotenv.load
-
-    # use rspec for tests
-    config.generators do |g|
-      g.test_framework :rspec
-    end
-
-    def self.activate
-      Dir.glob(File.join(File.dirname(__FILE__), '../../app/**/*_decorator*.rb')) do |c|
-        Rails.configuration.cache_classes ? require(c) : load(c)
-      end
-    end
-
-    config.to_prepare(&method(:activate).to_proc)
-  end
-end
