+++ lib/spree/retail/factories.rb
